refactor: change directory order
